other examples now present (still unproven)
I wrote down all of the unique examples that remained for the most part
(some of them, for example, were using an “or” vs using the equivalence
in nested ifs… and all we have is ifs, so I just did the “if” version,
etc) - haven’t attempted to prove them. Some will likely be simple,
others will expose more opportunities to refine the logic and improve
the TR typechecking proof tactics (eautoT, eautoUL, crushTR, etc…).